+++ CMakeLists.txt
@@ -32,5 +32,4 @@ include("cmake/thirdparty.cmake")
 include("cmake/shaders.cmake")
 
 add_subdirectory(common)
-add_subdirectory(samples)
-add_subdirectory(tasks)
+add_subdirectory(demo)
