+++ demo/Renderer.hpp
@@ -11,14 +11,21 @@
 #include "WorldRenderer.hpp"
 
 
+class ImGuiRenderer;
+
 using ResolutionProvider = fu2::unique_function<glm::uvec2() const>;
 
+/**
+ * This class encapsulates things that are very unlikely to change from one sample to another.
+ * E.g. initialization, frame delivery logic, window resizing, gui setup, etc.
+ */
 class Renderer
 {
 public:
   explicit Renderer(glm::uvec2 resolution);
   ~Renderer();
 
+  // Initializing all of rendering is a tricky multi-step dance
   void initVulkan(std::span<const char*> instance_extensions);
   void initFrameDelivery(vk::UniqueSurfaceKHR surface, ResolutionProvider res_provider);
   void recreateSwapchain(glm::uvec2 res);
@@ -30,12 +37,11 @@ public:
 
 private:
   ResolutionProvider resolutionProvider;
-
   std::unique_ptr<etna::Window> window;
   std::unique_ptr<etna::PerFrameCmdMgr> commandManager;
 
   glm::uvec2 resolution;
-  bool useVsync = true;
+  std::unique_ptr<ImGuiRenderer> guiRenderer;
 
   std::unique_ptr<WorldRenderer> worldRenderer;
 };
