+++ demo/shaders/simple_shadow.frag
