+++ resources/scenes/DamagedHelmet/LICENSE.md
@@ -1,0 +1,17 @@
+# LICENSE file for the model: Damaged Helmet
+
+All files in this directory tree are licensed as indicated below.
+
+* All files directly associated with the model including all text, image and binary files:
+
+  * [CC BY 4.0 International]("https://creativecommons.org/licenses/by/4.0/legalcode") [SPDX license identifier: "CC-BY-4.0"]
+
+  * [CC BY-NC 4.0 International]("https://creativecommons.org/licenses/by-nc/4.0/legalcode") [SPDX license identifier: "CC-BY-NC-4.0"]
+
+* This file and all other metadocumentation files including "metadata.json":
+
+  * [Creative Commons Attribtution 4.0 International]("https://creativecommons.org/licenses/by/4.0/legalcode") [SPDX license identifier: "CC-BY-4.0"]
+
+Full license text of these licenses are available at the links above
+
+#### Generated by modelmetadata
