--- samples/CMakeLists.txt
@@ -1,4 +1,0 @@
-include(${PROJECT_SOURCE_DIR}/cmake/common.cmake)
-
-add_subdirectory(shadowmap)
-add_subdirectory(simple_compute)
