--- samples/simple_compute/execute.cpp
@@ -1,17 +1,0 @@
-#include "simple_compute.h"
-
-#include <etna/Etna.hpp>
-
-
-void SimpleCompute::execute()
-{
-  setup();
-
-  auto cmdBuf = cmdMgr->start();
-
-  buildCommandBuffer(cmdBuf);
-
-  cmdMgr->submitAndWait(std::move(cmdBuf));
-
-  readback();
-}
