--- samples/simple_compute/shaders/simple.comp
@@ -1,31 +1,0 @@
-#version 430
-
-layout(local_size_x = 32) in;
-
-layout(push_constant) uniform params
-{
-  uint len;
-} pushConstant;
-
-layout(std430, binding = 0) buffer a
-{
-    float A[];
-};
-
-layout(std430, binding = 1) buffer b
-{
-    float B[];
-};
-
-layout(std430, binding = 2) buffer Sum
-{
-    float sum[];
-};
-
-void main()
-{
-    uint idx = gl_GlobalInvocationID.x;
-    if (idx < pushConstant.len) {
-        sum[idx] = A[idx] + B[idx];
-    }
-}
