--- tasks/CMakeLists.txt
@@ -1,5 +1,0 @@
-include(${PROJECT_SOURCE_DIR}/cmake/common.cmake)
-
-add_subdirectory(local_shadertoy1)
-add_subdirectory(local_shadertoy2)
-add_subdirectory(model_bakery)
