--- tasks/README.md
@@ -1,17 +1,0 @@
-# Задания
-
-Все задания представляют собой изолированные проекты.
-Смело меняйте их `CMakeLists.txt` как вам угодно, добавляйте новые файлы и папки, а также копируйте код семплов.
-
-Каждую домашку следует выполнять в **отдельной ветке** с названием, соответствующим папке задания, и создавать pull-request в мастер для проверки.
-
-Код из папки [common](/common/) тоже разрешается и приветствуется менять, однако следует помнить, что после проверки ветка будет вмерджена в мастер и изменения в `common` сделанные для одной домашки могут конфликтовать или просто "молча" сломать другую домашку.
-Будьте аккуратны, старайтесь не ломать совместимость с ещё не проверенными домашками!
-
- 1. [Shadertoy 1](shadertoy1/)
- 2. [Shadertoy 2](shadertoy2/)
- 3. [Локальный shadertoy 1](local_shadertoy1/)
- 4. [Захват кадра](capture/)
- 5. [Локальный shadertoy 2](local_shadertoy2/)
- 6. [Кадры в полёте](inflight_frames/)
- 7. [Запекалка моделей](model_bakery/)
