--- tasks/capture/README.md
@@ -1,32 +1,0 @@
-# Захват кадра
-
-В данном задании вам предстоит познакомиться с [RenderDoc](https://renderdoc.org/) и посмотреть на реальных примерах как выглядит традиционный графический пайплайн.
-
-## Перед началом
-
-1. Скачайте и установите [RenderDoc](https://renderdoc.org/).
-2. Выберите свою любимую 3D игру с хоть сколько-то интересной графикой, предпочтительно использующую Вулкан. Также подойдёт DirectX12, DirectX11 или OpenGL. При этом слишком старые игры нам не подойдут.
-
-Заметка: если у вас видеокарта от NVidia, вместо RenderDoc можно попробовать использовать [NSight](https://developer.nvidia.com/nsight-graphics).
-
-## Задание
-
-Захватите кадр из вашей любимой игры при помощи RenderDoc.
-Предпочтительно, чтобы в игре была интересная графика (майнкрафт без шейдеров -- не очень интересно) и чтобы эта игра использовала Vulkan.
-
-Иногда запустить игру напрямую из RenderDoc не получается из-за античитов.
-В таком случае попробуйте использовать возможность инъекции RenderDoc в уже запущенный процесс, или попробуйте использовать NSight.
-
-Просмотрите захват на предмет интересных техник и приёмов.
-В каком порядке и что рендерит игра?
-Как она рендерит геометрию сцены?
-Протыкайте все кнопочки что видите в рнедердоке, просмотрите все окошки.
-Что вам в них показывают?
-Попробуйте почитать байткод шейдеров и поредактировать их.
-
-Напишите короткий отчёт со скриншотами о ваших находках в произвольном формате.
-
-#### Заметка
-
-Делать пулл-реквест для этой задачи не нужно, достаточно прислать ваш мини-отчёт со скриншотами/картинками.
-
