--- tasks/local_shadertoy1/App.hpp
@@ -1,31 +1,0 @@
-#pragma once
-
-#include <etna/Window.hpp>
-#include <etna/PerFrameCmdMgr.hpp>
-#include <etna/ComputePipeline.hpp>
-#include <etna/Image.hpp>
-
-#include "wsi/OsWindowingManager.hpp"
-
-
-class App
-{
-public:
-  App();
-  ~App();
-
-  void run();
-
-private:
-  void drawFrame();
-
-private:
-  OsWindowingManager windowing;
-  std::unique_ptr<OsWindow> osWindow;
-
-  glm::uvec2 resolution;
-  bool useVsync;
-
-  std::unique_ptr<etna::Window> vkWindow;
-  std::unique_ptr<etna::PerFrameCmdMgr> commandManager;
-};
