--- tasks/local_shadertoy1/shaders/toy.comp
@@ -1,18 +1,0 @@
-#version 430
-
-layout(local_size_x = 32, local_size_y = 32) in;
-
-layout(binding = 0, rgba8) uniform image2D resultImage;
-
-
-void main()
-{
-  ivec2 uv = ivec2(gl_GlobalInvocationID.xy);
-
-  // TODO: Put your shadertoy code here!
-  // Simple gradient as a test.
-  vec3 color = vec3(vec2(uv) / vec2(1280, 720), 0);
-
-  if (uv.x < 1280 && uv.y < 720)
-    imageStore(resultImage, uv, vec4(color, 1));
-}
