--- tasks/model_bakery/CMakeLists.txt
@@ -1,5 +1,0 @@
-# Simple app to render models
-add_subdirectory(renderer)
-
-# 3D asset baker
-add_subdirectory(baker)
