--- tasks/model_bakery/baker/CMakeLists.txt
@@ -1,7 +1,0 @@
-
-add_executable(model_bakery_baker
-  main.cpp
-)
-
-target_link_libraries(model_bakery_baker
-  PRIVATE tinygltf)
