--- tasks/model_bakery/renderer/FramePacket.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <scene/Camera.hpp>
-
-
-struct FramePacket
-{
-  Camera mainCam;
-  float currentTime = 0;
-};
