--- tasks/shadertoy1/README.md
@@ -1,23 +1,0 @@
-# Shadertoy 1
-
-В данном задании вам предстоит развлечься и порисовать красивых картинок.
-
-## Перед началом
-
- 1. Убедитесь, что у вас есть аккаунт на сайте [shadertoy](https://www.shadertoy.com/), или заведите новый.
- 2. Вспомните материал лекции про SDF, их трассировку, вычислению освещения по Фонгу.
-
-## Задание
-
-Вдохновляясь другими шейдерами на shadertoy, напишите свой шейдер, трассирующий какую-то SDF с освещением.
-Строгих требований к шейдеру нет, но чем интереснее и красивее вы сделаете &mdash; тем лучше.
-
-#### Заметка
-
-Делать пулл-реквест для этой задачи не нужно, достаточно прислать ссылку на ваш шейдер-той.
-
-## Полезные материалы
-
- 1. https://www.khronos.org/opengl/wiki/OpenGL_Shading_Language &mdash; справочник по языку GLSL
- 2. https://iquilezles.org/ &mdash; сайт Иниго Килес со статьями и гайдами по шейдертою
- 3. https://steps3d.narod.ru/tutorials/sdf-3d-tutorial.html &mdash; введение в SDF от Борескова Алексея Викторовича
