--- tasks/shadertoy2/README.md
@@ -1,19 +1,0 @@
-# Shadertoy 2
-
-В данном задании мы продолжим развлекаться, улучшив шейдер из задания [shadertoy 1](/tasks/shadertoy1/).
-
-## Перед началом
-
- 1. Создайте **новый** шейдер на [shadertoy](https://www.shadertoy.com/) и скопируйте в него код из задания [shadertoy 1](/tasks/shadertoy1/).
- 2. Вспомните материал лекции про работу с текстурами.
-
-## Задание
-
- 1. Используя один из каналов (iChannelN) шейдертоя и встроенные кубмапы, добавьте в свой шейдер *скайбокс*.
- 2. Используя другой канал шейдертоя, натяните на свою SDF текстуру при помощи *трипланарной проекции*.
- 3. Cгенерируйте новым вспомогательным шейдером процедурную текстуру и положите её в канал текстуры геометрии.
-    Альтернативно, можете сгенерировать что-то более интересное и как-то иначе использовать в основном проходе.
-
-#### Заметка
-
-Делать пулл-реквест для этой задачи не нужно, достаточно прислать ссылку на ваш шейдер-той.
